Add stub Fuzz test and forge fmt

+++ uniswap-v2/src/UniswapV2ERC20.sol
@@ -6,7 +6,7 @@ import "./libraries/SafeMath.sol";
 
 // IUniswapV2ERC20 구현 (LP 토큰)
 contract UniswapV2ERC20 is IUniswapV2ERC20 {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     /*
         상태 변수 선언
@@ -14,25 +14,22 @@ contract UniswapV2ERC20 is IUniswapV2ERC20 {
     string public constant name = "Uniswap V2";
     string public constant symbol = "UNI-V2";
     uint8 public constant decimals = 18;
-    uint public totalSupply;
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    uint256 public totalSupply;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     bytes32 public DOMAIN_SEPARATOR;
-    bytes32 public constant PERMIT_TYPEHASH =
-        0x6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c9; // keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
-    mapping(address => uint) public nonces;
+    bytes32 public constant PERMIT_TYPEHASH = 0x6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c9; // keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
+    mapping(address => uint256) public nonces;
 
     constructor() {
-        uint chainId;
+        uint256 chainId;
         assembly {
             chainId := chainid()
         }
         DOMAIN_SEPARATOR = keccak256(
             abi.encode(
-                keccak256(
-                    "EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"
-                ),
+                keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
                 keccak256(bytes(name)), // name
                 keccak256(bytes("1")), // version
                 chainId, // chainId
@@ -46,27 +43,27 @@ contract UniswapV2ERC20 is IUniswapV2ERC20 {
      */
 
     // to에게 value만큼의 토큰을 발행
-    function _mint(address to, uint value) internal {
+    function _mint(address to, uint256 value) internal {
         totalSupply += value; // 총 발행량 증가
         balanceOf[to] += value; // to의 잔고 증가
         emit Transfer(address(0), to, value); // 이벤트 발생
     }
 
     // from의 토큰을 value만큼 소각
-    function _burn(address from, uint value) internal {
+    function _burn(address from, uint256 value) internal {
         balanceOf[from] -= value; // from의 잔고 감소
         totalSupply -= value; // 총 발행량 감소
         emit Transfer(from, address(0), value); // 이벤트 발생
     }
 
     // owner가 spender에게 value만큼의 토큰을 사용할 수 있도록 허락
-    function _approve(address owner, address spender, uint value) private {
+    function _approve(address owner, address spender, uint256 value) private {
         allowance[owner][spender] = value; // allowance 갱신
         emit Approval(owner, spender, value); // 이벤트 발생
     }
 
     // from에서 to로 value만큼의 토큰을 전송
-    function _transfer(address from, address to, uint value) private {
+    function _transfer(address from, address to, uint256 value) private {
         balanceOf[from] -= value; // from의 잔고 감소
         balanceOf[to] += value; // to의 잔고 증가
         emit Transfer(from, to, value); // 이벤트 발생
@@ -76,24 +73,20 @@ contract UniswapV2ERC20 is IUniswapV2ERC20 {
         공개 함수 선언
      */
     // spender가 value만큼의 토큰을 사용할 수 있도록 허락
-    function approve(address spender, uint value) external returns (bool) {
+    function approve(address spender, uint256 value) external returns (bool) {
         _approve(msg.sender, spender, value);
         return true;
     }
 
     // to에게 value만큼의 토큰을 전송
-    function transfer(address to, uint value) external returns (bool) {
+    function transfer(address to, uint256 value) external returns (bool) {
         _transfer(msg.sender, to, value);
         return true;
     }
 
     // from에서 to로 value만큼의 토큰을 전송
-    function transferFrom(
-        address from,
-        address to,
-        uint value
-    ) external returns (bool) {
-        if (allowance[from][msg.sender] != type(uint).max) {
+    function transferFrom(address from, address to, uint256 value) external returns (bool) {
+        if (allowance[from][msg.sender] != type(uint256).max) {
             // 무제한 허가가 아닌 경우 (허가량이 제한되어 있는 경우)
             allowance[from][msg.sender] -= value; // 허가량 감소
         }
@@ -106,8 +99,8 @@ contract UniswapV2ERC20 is IUniswapV2ERC20 {
     function permit(
         address owner, // 서명자
         address spender, // 허락 받는자
-        uint value, // 허락하는 토큰의 양
-        uint deadline, // 허락 기한 (type(uint).max: 무제한)
+        uint256 value, // 허락하는 토큰의 양
+        uint256 deadline, // 허락 기한 (type(uint).max: 무제한)
         uint8 v, // 서명의 v
         bytes32 r, // 서명의 r
         bytes32 s // 서명의 s
