+++ uniswap-v2/src/examples/SimpleFlashSwap.sol
@@ -36,11 +36,7 @@ contract SimpleFlashSwap is IUniswapV2Callee {
         return IERC20(token).transfer(owner, amount);
     }
 
-    function flashSwap(
-        address tokenBorrow,
-        address tokenRepay,
-        uint256 amount
-    ) external {
+    function flashSwap(address tokenBorrow, address tokenRepay, uint256 amount) external {
         if (tokenBorrow != token0 && tokenBorrow != token1) {
             revert InvalidToken();
         }
@@ -53,27 +49,18 @@ contract SimpleFlashSwap is IUniswapV2Callee {
             revert InvalidAmount();
         }
 
-        (uint amount0Out, uint amount1Out) = tokenBorrow == token0
-            ? (amount, uint(0))
-            : (uint(0), amount);
+        (uint256 amount0Out, uint256 amount1Out) = tokenBorrow == token0 ? (amount, uint256(0)) : (uint256(0), amount);
 
         bytes memory data = encodeData(tokenBorrow, tokenRepay, amount);
         pair.swap(amount0Out, amount1Out, address(this), data);
     }
 
-    function uniswapV2Call(
-        address sender,
-        uint amount0,
-        uint amount1,
-        bytes calldata data
-    ) external override {
+    function uniswapV2Call(address sender, uint256 amount0, uint256 amount1, bytes calldata data) external override {
         if (msg.sender != address(pair) || sender != address(this)) {
             revert InvalidPair();
         }
 
-        (address tokenBorrow, address tokenRepay, uint256 amount) = decodeData(
-            data
-        );
+        (address tokenBorrow, address tokenRepay, uint256 amount) = decodeData(data);
 
         if (tokenBorrow != token0 && tokenBorrow != token1) {
             revert InvalidToken();
@@ -87,7 +74,7 @@ contract SimpleFlashSwap is IUniswapV2Callee {
             revert InvalidAmount();
         }
 
-        uint repayAmount;
+        uint256 repayAmount;
 
         if (tokenBorrow == tokenRepay) {
             repayAmount = repayWithSameToken(amount);
@@ -98,17 +85,11 @@ contract SimpleFlashSwap is IUniswapV2Callee {
         IERC20(tokenRepay).transfer(address(pair), repayAmount);
     }
 
-    function encodeData(
-        address tokenBorrow,
-        address tokenRepay,
-        uint256 amount
-    ) public pure returns (bytes memory) {
+    function encodeData(address tokenBorrow, address tokenRepay, uint256 amount) public pure returns (bytes memory) {
         return abi.encode(tokenBorrow, tokenRepay, amount);
     }
 
-    function decodeData(
-        bytes calldata data
-    )
+    function decodeData(bytes calldata data)
         public
         pure
         returns (address tokenBorrow, address tokenRepay, uint256 amount)
@@ -116,23 +97,17 @@ contract SimpleFlashSwap is IUniswapV2Callee {
         return abi.decode(data, (address, address, uint256));
     }
     // about 0.3009% fee
-    function repayWithSameToken(
-        uint256 amount
-    ) public pure returns (uint256 repayAmount) {
+
+    function repayWithSameToken(uint256 amount) public pure returns (uint256 repayAmount) {
         repayAmount = (amount * 1000) / 997 + 1;
     }
 
-    function repayWithDifferentToken(
-        address tokenBorrow,
-        uint256 amount
-    ) public view returns (uint256 repayAmount) {
-        (uint256 reserve0, uint256 reserve1, ) = pair.getReserves();
-        (uint256 amount0Out, uint256 amount1Out) = tokenBorrow == token0
-            ? (amount, uint256(0))
-            : (uint256(0), amount);
+    function repayWithDifferentToken(address tokenBorrow, uint256 amount) public view returns (uint256 repayAmount) {
+        (uint256 reserve0, uint256 reserve1,) = pair.getReserves();
+        (uint256 amount0Out, uint256 amount1Out) = tokenBorrow == token0 ? (amount, uint256(0)) : (uint256(0), amount);
 
-        uint numerator;
-        uint denominator;
+        uint256 numerator;
+        uint256 denominator;
 
         if (amount0Out > 0) {
             numerator = amount0Out * reserve1 * 1000;
